Split the Header in a separate file and use NavLink instead of Link
